+++ js/api.js
@@ -28,7 +28,6 @@ const load = (route, errorText, method = Method.GET, body = null) =>
     });
 
 const getData = () => load(Route.GET_DATA, ErrorText.GET_DATA);
-
 const sendData = (body) => load(Route.SEND_DATA, ErrorText.SEND_DATA, Method.POST, body);
 
 export {getData, sendData};
