+++ js/load-comments.js
@@ -1,10 +1,19 @@
 const COMMENT_SHOW_COUNT = 5;
+
 const bigPicture = document.querySelector('.big-picture');
 const commentsLoader = bigPicture.querySelector('.comments-loader');
 const commentList = bigPicture.querySelector('.social__comments');
 const commentsCount = bigPicture.querySelector('.social__comment-count');
 const commentListItem = bigPicture.querySelector('.social__comment');
+
 commentListItem.classList.add('hidden');
+
+const showComments = (hiddenComments, count) => {
+  for (let i = 0; i < count; i++) {
+    hiddenComments[i].classList.remove('hidden');
+  }
+  commentsCount.textContent = `${commentList.children.length - commentList.querySelectorAll('.hidden').length} из ${commentList.children.length} комментариев`;
+};
 
 const loadComments = () => {
   const hiddenComments = commentList.querySelectorAll('.hidden');
@@ -15,14 +24,6 @@ const loadComments = () => {
     commentsLoader.classList.add('hidden');
   }
 };
-
-function showComments (hiddenComments, count) {
-  for (let i = 0; i < count; i++) {
-    hiddenComments[i].classList.remove('hidden');
-  }
-  commentsCount.textContent = `${commentList.children.length - commentList.querySelectorAll('.hidden').length} из ${commentList.children.length} комментариев`;
-}
-
 
 commentsLoader.addEventListener('click', loadComments);
