+++ js/main.js
@@ -1,6 +1,7 @@
 import {renderGallery} from './gallery.js';
 import {getData} from './api.js';
-import {showAlert, debounce} from './universal.js';
+import {showAlert} from './messages.js';
+import {debounce} from './debounce.js';
 import {loadLocalFile} from './local-picture.js';
 import {init, getFilter} from './filter.js';
 import {formSubmit} from './form.js';
